+++ sebastian/src/lib.rs
@@ -20,6 +20,7 @@ pub mod pie;
 pub mod quadrant;
 pub mod radar;
 pub mod render;
+pub mod sankey;
 pub mod sequence;
 pub mod state;
 pub mod svg;